+++ scripts/class_generators/generator.py
