+++ src/graphql/models/annotation_model.py
